pyproject.toml
[tool.poetry]
name = "openstates-scrapers"
version = "2022.7"
description = ""
authors = ["James Turk <dev@jamesturk.net>"]

[tool.poetry.dependencies]
python = "^3.9"
xlrd = "<2"
lxml = "^4.4"
pytz = "^2019.3"
feedparser = "^6.0"
python-dateutil = "^2.8"
spatula = "^0.9"
chardet = "^3.0"
suds-py3 = "^1.3"
requests = {version = "^2.22", extras = ["security"]}
mysqlclient = {version = "^1.4.6", optional=true}
SQLAlchemy = {version = "^1.3", optional=true}
cloudscraper = "^1.2.58"
paramiko = "^2.9.2"
cryptography = "^37.0.2"
openstates = "^6.17"
demjson3 = "^3.0.6"

[tool.poetry.extras]
california = ["mysqlclient", "SQLAlchemy"]

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^22"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
